refactor: Search 하위 모듈 3개를 app/functions/search.rb로 병합
- Search::VectorMath, ReciprocalRankFusion, MaximalMarginalRelevance
  세 모듈을 한 파일로 합쳐 파일 수 감소
- 테스트도 test/functions/search_test.rb로 병합
